First pass of networking support
Basic URL fetching works on Linux and PS3. Compiles on Windows
but is not working until we can initialize Winsock.

Move object watcher from lbshell into base.

Modify gyp files to reference googleurl/googleurl.gyp rather than
build/temp_gyp/googleurl.gyp, and third_party/openssl.

Disable net_resources actions for Cobalt- we don't have tools/grit,
and I don't think net_resources is needed.

Set use_native_http_stack variable default to 0.

Some DIAL compilation fixes for Linux- don't use file-scope
std::strings as they require global destructors to run.

Fix a few compiler warnings in under MSVC related to
signed/unsigned conversion and type narrowing.

Add actual_target_arch to filename_rules.gypi, to allow
building lbshell/src/*_win.cc on Windows.

Fix a few lint errors.

+++ base/threading/simple_thread.h
@@ -73,12 +73,13 @@ class BASE_EXPORT SimpleThread : public PlatformThread::Delegate {
     size_t stack_size() const { return stack_size_; }
 
     // Thread priority
-    void set_priority(size_t priority) { priority_ = priority; }
-    size_t priority() const { return priority_; }
+    void set_priority(int priority) { priority_ = priority; }
+    int priority() const { return priority_; }
 
     // Core affinity
-    void set_affinity(size_t affinity) { affinity_ = affinity; }
-    size_t affinity() const { return affinity_; }
+    void set_affinity(int affinity) { affinity_ = affinity; }
+    int affinity() const { return affinity_; }
+
    private:
     size_t stack_size_;
     int priority_;
